Move is_local to _container_io_manager
