+++ modal/functions.py
@@ -32,6 +32,7 @@ from synchronicity.exceptions import UserCodeException
 
 from modal_proto import api_grpc, api_pb2
 
+from ._container_io_manager import is_local
 from ._location import parse_cloud_provider
 from ._output import OutputManager
 from ._pty import get_pty_info
@@ -53,7 +54,6 @@ from ._utils.blob_utils import (
 from ._utils.function_utils import FunctionInfo, _stream_function_call_data, get_referred_objects, is_async
 from ._utils.grpc_utils import retry_transient_errors
 from ._utils.mount_utils import validate_mount_points, validate_volumes
-from .app import is_local
 from .call_graph import InputInfo, _reconstruct_call_graph
 from .client import _Client
 from .cloud_bucket_mount import _CloudBucketMount, cloud_bucket_mounts_to_proto
